Make time calculations always long long.
Previously pgpool assumed that time_t to be a simple long. This causes
a lot of compile time warnings on certain systems, for example
OpenBSD because on the system time_t is __int64, which results in long
long. This commit upcasts such calculations to long long to avoid the
issue.

This way times can't get truncated and for the places where time_t is
actually used as a time and not a time diff this would allow pgpool to
keep working correctly post Y2038 on 64 bit clean time_t systems
(e.g. i386 OpenBSD).

Moreover, json_get_long_value_for_key is changed to
json_get_llong_value_for_key and changed the parameter to long long.
This makes it more in line _json_value's integer, which is defined as
int64. This should also give 32 bit platforms proper retrieval of the
max value of an integer and may or may not solve some weird integer
overflow issues.

Backpatch for 4.6 was rebased by Gyorgy Sarvari.

Author: Martijn van Duren <pgpool@list.imperialat.at>
Backpatch-through: v4.6
Discussion: https://www.pgpool.net/pipermail/pgpool-hackers/2025-May/004584.html
Discussion: https://www.postgresql.org/message-id/20251003.211957.2067537305399895611.ishii%40postgresql.org
Discussion: https://github.com/pgpool/pgpool2/pull/128

+++ src/watchdog/wd_heartbeat.c
@@ -850,8 +850,8 @@ packet_to_string_hb(WdHbPacket *pkt, char *str, int maxlen)
 {
 	int			len;
 
-	len = snprintf(str, maxlen, "tv_sec=%ld tv_usec=%ld from=%s",
-				   pkt->send_time.tv_sec, pkt->send_time.tv_usec, pkt->from);
+	len = snprintf(str, maxlen, "tv_sec=%lld tv_usec=%lld from=%s",
+				   (long long)pkt->send_time.tv_sec, (long long)pkt->send_time.tv_usec, pkt->from);
 
 	return len;
 }
